youtube_trends/requirements.txt
apify-client==1.8.1
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
celery==5.4.0
redis==5.0.6
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==24.1.0
plotly==5.22.0
sendgrid==6.11.0
slack-sdk==3.31.0
pytest==8.2.2
pytest-asyncio==0.24.0
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2